pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.30.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
beautifulsoup4 = ">=4.12.0"
lxml = ">=4.9.0"
requests =">=2.31.0"
alembic = "^1.13.0"
pandas = "^2.1.4"
python-dateutil = "^2.8.2"
plotly = "^5.18.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
httpx = ">=0.27,<0.29"
mistralai = "^1.0.0"
# API REST
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
# Authentication
supabase = "^2.3.0"
# Barcode scanning
pyzbar = {version = "^0.1.9", optional = true}
opencv-python-headless = {version = "^4.9.0", optional = true}
streamlit-webrtc = {version = "^0.47.0", optional = true}
# PDF generation
reportlab = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.2"
black = "^23.12.1"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abstractmethod",
]
precision = 2
skip_covered = false
skip_empty = true
sort = "Name"

[tool.coverage.html]
directory = "htmlcov"

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__", "**/.*"]
typeCheckingMode = "strict"
reportGeneralTypeIssues = "error"
reportPrivateUsage = "warning"
reportOptionalMemberAccess = "warning"
reportOptionalSubscript = "warning"
